requirements.txt
Flask==3.0.3
google==3.0.0
google-api-core==2.19.1
google-auth==2.30.0
google-cloud-core==2.4.1
google-cloud-storage==2.17.0
google-crc32c==1.5.0
google-resumable-media==2.7.1
googleapis-common-protos==1.63.2
hachoir==3.3.0
Jinja2==3.1.4
pillow==10.3.0
python-dotenv==1.0.1
requests==2.32.3
Werkzeug==3.0.3
